refactor(workspace): extract kanban surface
