+++ crates/gwt/src/embedded_server.rs
@@ -131,6 +131,10 @@ impl EmbeddedServer {
                 get(embedded_web::board_surface_js_handler),
             )
             .route(
+                "/workspace-kanban-surface.js",
+                get(embedded_web::workspace_kanban_surface_js_handler),
+            )
+            .route(
                 "/assets/xterm/xterm.mjs",
                 get(embedded_web::xterm_js_handler),
             )
