+++ crates/gwt/src/embedded_web.rs
@@ -27,6 +27,10 @@ pub fn board_surface_js() -> &'static str {
     include_str!("../web/board-surface.js")
 }
 
+pub fn workspace_kanban_surface_js() -> &'static str {
+    include_str!("../web/workspace-kanban-surface.js")
+}
+
 pub fn xterm_js() -> &'static str {
     include_str!("../web/vendor/xterm/xterm.mjs")
 }
@@ -129,6 +133,13 @@ pub async fn board_surface_js_handler() -> impl IntoResponse {
     (
         [(header::CONTENT_TYPE, "text/javascript; charset=utf-8")],
         board_surface_js(),
+    )
+}
+
+pub async fn workspace_kanban_surface_js_handler() -> impl IntoResponse {
+    (
+        [(header::CONTENT_TYPE, "text/javascript; charset=utf-8")],
+        workspace_kanban_surface_js(),
     )
 }
 
@@ -244,11 +255,12 @@ pub async fn font_jetbrains_mono_handler() -> impl IntoResponse {
 mod tests {
     use super::{
         app_js, board_surface_js, branch_cleanup_modal_js, index_html, window_docking_js,
-        xterm_css, xterm_fit_js, xterm_js,
+        workspace_kanban_surface_js, xterm_css, xterm_fit_js, xterm_js,
     };
     use super::{
         app_js_handler, board_surface_js_handler, branch_cleanup_modal_js_handler,
-        window_docking_js_handler, xterm_css_handler, xterm_fit_js_handler, xterm_js_handler,
+        window_docking_js_handler, workspace_kanban_surface_js_handler, xterm_css_handler,
+        xterm_fit_js_handler, xterm_js_handler,
     };
     // SPEC-2356 — Operator Design System modules.
     use super::{focus_trap_js, hotkey_js, operator_shell_js, theme_manager_js, theme_toggle_js};
@@ -263,7 +275,9 @@ mod tests {
             "\n",
             include_str!("../web/app.js"),
             "\n",
-            include_str!("../web/board-surface.js")
+            include_str!("../web/board-surface.js"),
+            "\n",
+            include_str!("../web/workspace-kanban-surface.js")
         )
     }
 
@@ -547,6 +561,7 @@ mod tests {
         assert!(!focus_trap_js().is_empty());
         assert!(!window_docking_js().is_empty());
         assert!(!board_surface_js().is_empty());
+        assert!(!workspace_kanban_surface_js().is_empty());
         assert!(theme_manager_js().contains("createThemeManager"));
         assert!(theme_toggle_js().contains("wireThemeToggle"));
         assert!(hotkey_js().contains("createHotkeyManager"));
@@ -554,6 +569,7 @@ mod tests {
         assert!(focus_trap_js().contains("createFocusTrap"));
         assert!(window_docking_js().contains("findTitlebarDockTarget"));
         assert!(board_surface_js().contains("boardEntryMentionsSelf"));
+        assert!(workspace_kanban_surface_js().contains("createWorkspaceKanbanSurface"));
     }
 
     #[tokio::test]
@@ -572,6 +588,15 @@ mod tests {
         );
         assert_eq!(
             branch_cleanup_modal_js_handler()
+                .await
+                .into_response()
+                .headers()
+                .get(header::CONTENT_TYPE)
+                .unwrap(),
+            js,
+        );
+        assert_eq!(
+            workspace_kanban_surface_js_handler()
                 .await
                 .into_response()
                 .headers()
@@ -1152,6 +1177,7 @@ mod tests {
             "/branch-cleanup-modal.js",
             "/migration-modal.js",
             "/board-surface.js",
+            "/workspace-kanban-surface.js",
         ] {
             assert!(
                 js.contains(module_path),
