+++ crates/gwt/web/__tests__/operator-chrome-structure.test.mjs
@@ -1,6 +1,6 @@
 import { test } from "node:test";
 import assert from "node:assert/strict";
-import { readFileSync } from "node:fs";
+import { existsSync, readFileSync } from "node:fs";
 import { fileURLToPath } from "node:url";
 import { dirname, resolve } from "node:path";
 import { parseHTML } from "linkedom";
@@ -13,6 +13,11 @@ const operatorShellSource = readFileSync(resolve(here, "../operator-shell.js"), 
 const appSource = readFileSync(resolve(here, "../app.js"), "utf8");
 const branchCleanupSource = readFileSync(resolve(here, "../branch-cleanup-modal.js"), "utf8");
 const windowDockingSource = readFileSync(resolve(here, "../window-docking.js"), "utf8");
+const workspaceKanbanPath = resolve(here, "../workspace-kanban-surface.js");
+const workspaceKanbanSource = existsSync(workspaceKanbanPath)
+  ? readFileSync(workspaceKanbanPath, "utf8")
+  : "";
+const workspaceKanbanCombinedSource = `${appSource}\n${workspaceKanbanSource}`;
 const typographySource = readFileSync(resolve(here, "../styles/typography.css"), "utf8");
 const inlineStyle = html.match(/<style>([\s\S]*?)<\/style>/)?.[1] || "";
 
@@ -331,37 +336,64 @@ test("Workspace Overview is separate from live-only Active Work", () => {
 });
 
 test("Workspace Overview uses the shared full-window Kanban + detail layout", () => {
+  assert.ok(
+    workspaceKanbanSource.length > 0,
+    "expected Workspace Kanban renderer to live in workspace-kanban-surface.js",
+  );
+  assert.match(
+    appSource,
+    /from\s+"\/workspace-kanban-surface\.js"/,
+    "expected app.js to import the Workspace Kanban surface module",
+  );
   assert.match(
     appSource,
     /presetSurface\(preset\)[\s\S]+preset\s*===\s*"workspace"[\s\S]+return\s+"workspace"/,
     "expected Workspace to be a first-class window surface",
   );
   assert.match(
-    appSource,
+    workspaceKanbanCombinedSource,
     /workspace-kanban-root[\s\S]+workspace-split[\s\S]+kanban-shell/,
     "expected Workspace Overview to share the split Kanban shell",
   );
   for (const column of ["Active", "Suspended", "Completed"]) {
     assert.match(
-      appSource,
+      workspaceKanbanCombinedSource,
       new RegExp(`workspace-column-name">${column}`),
       `expected Workspace Kanban to include ${column} column`,
     );
   }
   assert.match(
-    appSource,
+    workspaceKanbanCombinedSource,
     /workspace-kanban-detail-pane/,
     "expected Workspace Kanban to keep selected Workspace detail on the right",
   );
   assert.match(
-    appSource,
+    workspaceKanbanSource,
     /function\s+workspaceCardsFromProjection\([^)]*\)[\s\S]+journal_entries/,
     "expected Workspace Kanban to render Workspace journal entries from active_work_projection",
   );
   assert.match(
-    appSource,
+    workspaceKanbanSource,
     /function\s+resumeWorkspaceCard\([^)]*\)[\s\S]+open_active_work_launch_wizard[\s\S]+open_start_work/,
     "expected every Workspace card Resume action to prefer branch resume and fall back to Start Work",
+  );
+  assert.doesNotMatch(
+    appSource,
+    /function\s+(workspaceCardsFromProjection|renderWorkspaceKanbanCard|renderWorkspaceKanbanDetail)\(/,
+    "Workspace Kanban rendering internals should not remain in app.js",
+  );
+});
+
+test("Workspace Overview legacy drawer scaffold is retired", () => {
+  assert.doesNotMatch(
+    html,
+    /workspace-overview-drawer/,
+    "Workspace Overview should no longer ship an unused drawer scaffold",
+  );
+  assert.doesNotMatch(
+    appSource,
+    /closeWorkspaceOverview|renderWorkspaceOverview|workspaceOverviewFocusTrapRelease/,
+    "Workspace Overview drawer lifecycle code should not remain in app.js",
   );
 });
