+++ crates/gwt/web/index.html
@@ -3685,37 +3685,6 @@
         <div class="op-drawer__body" id="kanban-drawer-body"></div>
         <footer class="op-drawer__footer" id="kanban-drawer-footer"></footer>
       </aside>
-      <div
-        class="op-drawer-backdrop"
-        id="workspace-overview-drawer-backdrop"
-        data-open="false"
-        hidden
-      ></div>
-      <aside
-        class="op-drawer workspace-overview-drawer"
-        id="workspace-overview-drawer"
-        role="dialog"
-        aria-modal="true"
-        aria-labelledby="workspace-overview-title"
-        data-open="false"
-        hidden
-        tabindex="-1"
-      >
-        <header class="op-drawer__header">
-          <h2
-            class="op-drawer__title"
-            id="workspace-overview-title"
-          >Workspace Overview</h2>
-          <button
-            type="button"
-            class="op-drawer__close"
-            id="workspace-overview-close"
-            aria-label="Close Workspace Overview"
-          >×</button>
-        </header>
-        <div class="op-drawer__body" id="workspace-overview-body"></div>
-        <footer class="op-drawer__footer" id="workspace-overview-footer"></footer>
-      </aside>
     </div>
 
     <script type="module" src="/app.js"></script>
